fix an altloc port parsing bug

+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/altlocs/AlternateLocationFactoryImpl.java
@@ -186,7 +186,7 @@ public class AlternateLocationFactoryImpl implements AlternateLocationFactory {
             if(location.length() < port+1)
                 throw new IOException("invalid location: " + location);
             try {
-                port = Short.parseShort(location.substring(port+1));
+                port = Integer.parseInt(location.substring(port+1));
             } catch(NumberFormatException nfe) {
                 throw new IOException("invalid location: " + location);
             }
